reventless: add Util.Pulumi.Output.Async module
The module's make function returns a tuple of a promise wrapped as an output and a function to resolve that promise with a given output.
This pattern enables us to depend on some resource-outputs (in runtime-code!) before they are actually created.
If this is used in deploy-time code, this could potentially introduce dependency-cycles.